requirements.txt
# DSA Content Engine v2 — Python dependencies
# Install: pip install -r requirements.txt

# ── AI / LLM ──────────────────────────────────────────────────────────
google-generativeai>=0.7.0        # Gemini Flash — scripts, copy, planning
anthropic>=0.34.0                  # Claude Haiku — copy quality review

# ── Animation ─────────────────────────────────────────────────────────
manim>=0.18.0                      # Core animation engine
manim-dsa>=0.4.0                   # DSA-specific: MArray, MStack, MGraph

# ── TTS ───────────────────────────────────────────────────────────────
kokoro>=0.9.2                      # AI voice (Apache 2.0, CPU-capable)
soundfile>=0.12.1                  # WAV read/write
numpy>=1.26.0                      # Audio array manipulation

# ── Image / Video ─────────────────────────────────────────────────────
Pillow>=10.0.0                     # Carousel slides + thumbnails
pygments>=2.17.0                   # Syntax highlighting in code slides

# ── HTTP / APIs ───────────────────────────────────────────────────────
requests>=2.31.0                   # LinkedIn, Instagram, YouTube APIs

# ── Telegram bot ──────────────────────────────────────────────────────
python-telegram-bot>=20.0          # Preview bot

# ── Dev / test ────────────────────────────────────────────────────────
pytest>=7.4.0
